Added the ability to create user-defined HTML styles, and for yaYUL/yaLEMAP to read a default file defining the style at startup. This makes it easy for annotation styles in the HTML assembly listings to be rather fancy without having to hard-code the style into the program or to define it in the AGC/AGS source files.  The default HTML style I created (and the graphic it uses) have been added to the svn project.
I removed all of the sample HTML that had been inserted 
into source files so far, but added some actual 
annotations (not many!) that may be of permanent value 
in the A11 LM's PINBALL_GAME_BUTTONS_AND_LIGHTS.s.

There is still a bug in which the assembler fails if 
styles are changed on the fly, so it is best at the 
moment to stick with the default HTML style throughout 
any given assembly listing.